PEP-compliant formatting
